pyproject.toml
[tool.poetry]
name = "mgy-bot"
version = "0.1.0"
description = ""
authors = ["Lucas Souto <lucassouto@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
dnspython = "^2.4.2"
pynacl = "^1.5.0"
async-timeout = "^4.0.3"
youtube-dl-server = "^0.3"
youtube-dl = "^2021.12.17"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
requests = "^2.31.0"
requests-html = "^0.10.0"
psycopg2-binary = "^2.9.9"
urllib3 = "^2.0.7"
google-api-python-client = "^2.107.0"
google-api-python-client-stubs = "^1.20.1"
spotipy = "^2.23.0"
python-dotenv = "^1.0.0"
pycord = "^0.1.1"
yt-dlp = "^2023.10.13"
sqlalchemy = {version="^2.0.23", extras=["asyncio"]}
asyncpg = "^0.29.0"
alembic = "^1.12.1"
python-decouple = "^3.8"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.5"
isort = "^5.12.0"


[tool.ruff]
select = ["RUF", "I", "PL", "F", "COM", "UP", "DJ", "T10", "T20", "DTZ", "SIM", "TID", "PTH", "ERA", "TRY"]
ignore = ["COM812", "PLR2004", "PLR0911", "PLR0913", "PLR0915", "DJ001", "DJ001", "DJ008", "TRY003", "PLR0912", "PLC1901", "RUF012", "PLW0603"]
line-length = 120
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"settings.py" = ["F403", "F404", "F405"]

[tool.ruff.isort]
known-first-party = [
    "cogs",
    "games",
    "models",
    "repositories",
    "utils",
    "bot",
    "database",
    "dependencies",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
